+++ apps/uniapp/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.11](https://gitee.com/newgateway/vtj/compare/vtj-project-uniapp@0.9.10...vtj-project-uniapp@0.9.11) (2025-01-09)
+
+**Note:** Version bump only for package vtj-project-uniapp
+
+
+
+
+
+## [0.9.10](https://gitee.com/newgateway/vtj/compare/vtj-project-uniapp@0.9.9...vtj-project-uniapp@0.9.10) (2025-01-09)
+
+**Note:** Version bump only for package vtj-project-uniapp
+
+
+
+
+
 ## [0.9.9](https://gitee.com/newgateway/vtj/compare/vtj-project-uniapp@0.9.8...vtj-project-uniapp@0.9.9) (2025-01-08)
 
 **Note:** Version bump only for package vtj-project-uniapp